Day/night cycle.

+++ VoxEngine/res/Shaders/Chunk/nonAlignedFace.vert
@@ -16,7 +16,7 @@ layout(binding = 0) restrict readonly buffer chunkPositionSSBO
 uniform mat4 view;
 uniform mat4 projection;
 uniform ivec3 cameraChunkPosition;
-uniform int skyLightSub = 0; // TODO: Maybe make it float for smooth sky light transition
+uniform float skyLightSub = 0;
 
 out vec2 uv;
 flat out uint textureID;
@@ -28,14 +28,14 @@ const float INV_LIGHT_SCALE = 1.0 / 15.0;
 const float INV_AO_SCALE = 1.0 / 3.0;
 const float AO_RANGE = 0.9; // [0; 1]
 
-uint getBlockLight(uint data, int i)
+int getBlockLight(uint data, int i)
 {
-    return bitfieldExtract(data, i * 8, 4);
+    return int(bitfieldExtract(data, i * 8, 4));
 }
 
-uint getSkyLight(uint data, int i)
+int getSkyLight(uint data, int i)
 {
-    return bitfieldExtract(data, i * 8 + 4, 4);
+    return int(bitfieldExtract(data, i * 8 + 4, 4));
 }
 
 uint getAO(uint data, int i)
@@ -66,23 +66,23 @@ void main()
     v = bitfieldExtract(instanceVsAndTextureId, gl_VertexID * 5, 5);
     textureID = bitfieldExtract(instanceVsAndTextureId, 20, 12);
 
-    blockLight[0] = int(getBlockLight(instanceLight.x, 0));
-    blockLight[1] = int(getBlockLight(instanceLight.x, 1));
-    blockLight[2] = int(getBlockLight(instanceLight.x, 2));
-    blockLight[3] = int(getBlockLight(instanceLight.x, 3));
-    blockLight[4] = int(getBlockLight(instanceLight.y, 0));
-    blockLight[5] = int(getBlockLight(instanceLight.y, 1));
-    blockLight[6] = int(getBlockLight(instanceLight.y, 2));
-    blockLight[7] = int(getBlockLight(instanceLight.y, 3));
+    blockLight[0] = getBlockLight(instanceLight.x, 0);
+    blockLight[1] = getBlockLight(instanceLight.x, 1);
+    blockLight[2] = getBlockLight(instanceLight.x, 2);
+    blockLight[3] = getBlockLight(instanceLight.x, 3);
+    blockLight[4] = getBlockLight(instanceLight.y, 0);
+    blockLight[5] = getBlockLight(instanceLight.y, 1);
+    blockLight[6] = getBlockLight(instanceLight.y, 2);
+    blockLight[7] = getBlockLight(instanceLight.y, 3);
 
-    skyLight[0] = int(getSkyLight(instanceLight.x, 0)) - skyLightSub;
-    skyLight[1] = int(getSkyLight(instanceLight.x, 1)) - skyLightSub;
-    skyLight[2] = int(getSkyLight(instanceLight.x, 2)) - skyLightSub;
-    skyLight[3] = int(getSkyLight(instanceLight.x, 3)) - skyLightSub;
-    skyLight[4] = int(getSkyLight(instanceLight.y, 0)) - skyLightSub;
-    skyLight[5] = int(getSkyLight(instanceLight.y, 1)) - skyLightSub;
-    skyLight[6] = int(getSkyLight(instanceLight.y, 2)) - skyLightSub;
-    skyLight[7] = int(getSkyLight(instanceLight.y, 3)) - skyLightSub;
+    skyLight[0] = int(getSkyLight(instanceLight.x, 0) - skyLightSub);
+    skyLight[1] = int(getSkyLight(instanceLight.x, 1) - skyLightSub);
+    skyLight[2] = int(getSkyLight(instanceLight.x, 2) - skyLightSub);
+    skyLight[3] = int(getSkyLight(instanceLight.x, 3) - skyLightSub);
+    skyLight[4] = int(getSkyLight(instanceLight.y, 0) - skyLightSub);
+    skyLight[5] = int(getSkyLight(instanceLight.y, 1) - skyLightSub);
+    skyLight[6] = int(getSkyLight(instanceLight.y, 2) - skyLightSub);
+    skyLight[7] = int(getSkyLight(instanceLight.y, 3) - skyLightSub);
 
     faceAO[0] = 1.0 - AO_RANGE + float(getAO(instanceAO, 0)) * INV_AO_SCALE * AO_RANGE;
     faceAO[1] = 1.0 - AO_RANGE + float(getAO(instanceAO, 1)) * INV_AO_SCALE * AO_RANGE;
